clear app cache page
