+++ packages/shared/src/routes/setting.ts
@@ -1,6 +1,7 @@
 export enum EModalSettingRoutes {
   SettingListModal = 'SettingListModal',
   SettingCurrencyModal = 'SettingCurrencyModal',
+  SettingClearAppCache = 'SettingClearAppCache',
   SettingAccountDerivationModal = 'SettingAccountDerivationModal',
   SettingSpendUTXOModal = 'SettingSpendUTXOModal',
   SettingCustomRPC = 'SettingCustomRPC',
@@ -20,6 +21,7 @@ export enum EModalSettingRoutes {
 export type IModalSettingParamList = {
   [EModalSettingRoutes.SettingListModal]: { flag?: string } | undefined;
   [EModalSettingRoutes.SettingCurrencyModal]: undefined;
+  [EModalSettingRoutes.SettingClearAppCache]: undefined;
   [EModalSettingRoutes.SettingAccountDerivationModal]: undefined;
   [EModalSettingRoutes.SettingSpendUTXOModal]: undefined;
   [EModalSettingRoutes.SettingCustomRPC]: undefined;
